:recycle: [REFACTOR] (#8) Long으로 수정

+++ app/src/main/java/com/sopt/dive/data/repository/UserRepository.kt
@@ -4,7 +4,7 @@ import com.sopt.dive.data.dto.response.UserDetailResponseDto
 import com.sopt.dive.data.service.UserService
 
 class UserRepository(private val userService: UserService) {
-    suspend fun getUser(userId: Int): Result<UserDetailResponseDto> = runCatching {
+    suspend fun getUser(userId: Long): Result<UserDetailResponseDto> = runCatching {
         val response = userService.getUser(userId)
 
         if (response.success) {
